requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
gunicorn==19.9.0
idna==2.7
mysqlclient==1.4.2.post1
Pillow==5.3.0
pytz==2018.9
requests==2.20.1
urllib3==1.24.1
